Changed listing_detail,place_bid functions and listing_detail.html file.
In listing_detail function added bids_count and bid_obj variables.

In place_bid function added check for highest_bid variable value exist.

In listing_detail.html added row with the numbers of bids for listing
and checked whether the highest bid belongs to the current user, and
if so,the site displays a corresponding message.